Add core
