+++ lazyr/__version__.py
@@ -1,3 +1,4 @@
+"""Version file."""
 VERSION = (0, 0, 7)
 
 __version__ = ".".join(map(str, VERSION))
